Offload legacy app patching stuff to utils/legacy.py to limit the size of app.py
